Created the HtmlHeadHelper by pulling methods out of BlacklightHelper.  render_head_content() now includes content in the content_for(:head) blocks
